backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pillow==10.4.0
opencv-python==4.10.0.84
python-multipart==0.0.9
sentence-transformers
torch
torchvision
